% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{abovecanopytemp}
\alias{abovecanopytemp}
\title{Calculates temperature above canopy}
\usage{
abovecanopytemp(
  tz,
  uz,
  zu,
  zo,
  H,
  hgt,
  PAI,
  zm0 = 0.004,
  pk = 101.3,
  psi_h = 0
)
}
\arguments{
\item{tz}{temperature at height \code{zu} above the canopy (deg C)}

\item{uz}{wind speed at height \code{zu} above the canopy (m / s)}

\item{zu}{height of \code{tz} and \code{uz} (m)}

\item{zo}{height (m) for which temperature is required (see details)}

\item{H}{sensible heat flux density (W / m^2). See details.}

\item{hgt}{height of the canopy (m)}

\item{PAI}{total plant area index of the canopy. Used to calculate roughness lengths.}

\item{zm0}{roughness length governing momentum transfer of ground vegetation}

\item{pk}{atmospheric pressure (kPa)}

\item{psi_h}{diabatic correction factor for heat transfer}
}
\description{
calaculates the temperature at the top of the canopy based
on the standard logarithmic height profile
}
\details{
Estimation of \code{H} requires estimation of temperature,
so most either be derived by iteration of taken form the previous timestep.
\code{H} is given by the net energy balance equation: \code{H} = \code{Rabs} - \code{Rem} - \code{L} - \code{G}
where \code{Rabs} is absorbed radiation, \code{Rem} emitted radiation, \code{L} Latent heat
exchange and \code{G} ground heat flux. This function is not valid for temperatures
below canopy so \code{hgt} must be lower than \code{zi}.
}
\examples{
abovecanopytemp(11, 2, 2, 1.5, 500, 1, 3)
abovecanopytemp(11, 2, 2, 0.5, 500, 0.25, 3)
}
