% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dailyprecip}
\alias{dailyprecip}
\title{A vector of daily precipitation}
\format{An object of class \code{numeric} of length 365.}
\usage{
dailyprecip
}
\description{
A vector of daily precipitation (mm) in 1995 at Camborne weather station, Cornwall (50.2178N, 5.32656W)
}
\keyword{datasets}
