% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{.runmodelsnow}
\alias{.runmodelsnow}
\title{Internal function for running model with snow}
\usage{
.runmodelsnow(climdata, vegp, soilp, nmrout, reqhgt, lat, long,
  metopen = TRUE, windhgt = 2)
}
\description{
Internal function for running model with snow
}
