% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{.snowsort}
\alias{.snowsort}
\title{internal function to sort out vegetation parameters for snow}
\usage{
.snowsort(vegp, snow, i)
}
\description{
internal function to sort out vegetation parameters for snow
}
