% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{.snowtempf}
\alias{.snowtempf}
\title{Internation function for computing snow temperature}
\usage{
.snowtempf(snowdepth, snow, reqhgt)
}
\description{
Internation function for computing snow temperature
}
