% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{.windprofile}
\alias{.windprofile}
\title{Internal function for calculating lead absorbed radiation on vector}
\usage{
.windprofile(ui, zi, zo, a = 2, hgt, PAI, psi_m = 0, hgtg = 0.05 *
  hgt, zm0 = 0.004)
}
\description{
Internal function for calculating lead absorbed radiation on vector
}
