% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{leafabs}
\alias{leafabs}
\title{Calculates radiation absorbed by leaf}
\usage{
leafabs(
  Rsw,
  tme,
  tair,
  tground,
  lat,
  long,
  PAI,
  pLAI,
  x,
  refls,
  refw,
  refg,
  vegem,
  skyem,
  dp = NA,
  merid = round(long/15, 0) * 15,
  dst = 0,
  clump = 0
)
}
\arguments{
\item{Rsw}{Incoming shortwave radiation (W / m2)}

\item{tme}{POSIXlt object of time(s)}

\item{tair}{air temperature (deg C)}

\item{tground}{ground temperature (deg C)}

\item{lat}{latitude (decimal degrees)}

\item{long}{longitude (decimal degrees)}

\item{PAI}{Vector of plant area indices for each canopy layer}

\item{pLAI}{Proportion of plant area that is green vegetation}

\item{x}{the ratio of vertical to horizontal projections of leaf foliage}

\item{refls}{reflectivity of green vegetation to shortwave radiation}

\item{refw}{reflectivity of woody vegetation to shortwave radiation}

\item{refg}{reflectivity of ground to shortwave radiation}

\item{vegem}{emissivity of vegetation}

\item{skyem}{sky emissivity}

\item{dp}{proportion of \code{Rsw} that is diffuse radiation. If not provided, then calculated using \code{\link[microctools:difprop]{microctools::difprop()}}}

\item{merid}{optional longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{optional value representing the time difference from the timezone meridian}

\item{clump}{clumpiness factor for canopy (0-1)}
}
\value{
a list of with the following components:

\code{aRsw} absorbed Shortwave radiation (W / m2)

\code{aRlw} absorbed longwave radiation (W / m2)

\code{ref} mean area-wighted reflectivity of vegetation (green and woody)
}
\description{
Calculates the flux density of radiation absorbed by leaf.
}
