% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{leafem}
\alias{leafem}
\title{Calculates radiation emitted by leaf}
\usage{
leafem(tc, vegem)
}
\arguments{
\item{tc}{temperature (deg C)}

\item{vegem}{emissivity of leaf}
}
\value{
Flux density of radiation emitted by leaf (W / m2)
}
\description{
Calculates the flux density of radiation emitted by the leaf.
}
