% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{leaftemp}
\alias{leaftemp}
\title{Calculates leaf temperature of canopy layers}
\usage{
leaftemp(tair, relhum, pk, timestep, gt, gha, gv, Rabs, previn, vegp,
  soilp, theta)
}
\arguments{
\item{tair}{air temperature at two metres above canopy (deg C)}

\item{relhum}{relative humidity at two metres above canopy (Percentage)}

\item{pk}{air pressure at two metres above canopy (kPa)}

\item{timestep}{duration of time step (s)}

\item{gt}{heat conductance by turbulent convection (mol / m^2 / s)}

\item{gha}{leaf-air heat conductance (mol / m^2 / s)}

\item{gv}{leaf-air vapour conductance (mol / m^2 / s)}

\item{Rabs}{Flux density of absorbed radiation as returned by \code{\link[=leafabs]{leafabs()}} (W/m2)}

\item{previn}{list of values from previous timestep}

\item{vegp}{list of vegetation paramaters (see e.g. \code{vegparams} dataset)}

\item{soilp}{list Soil paramaters (see e.g. \code{soilparams} dataset)}

\item{theta}{volumetric soil moisture fraction of top soil layer (m^3 / m^3)}
}
\value{
a list with the following elements:

\code{tn} air temperature of each layer (deg C)

\code{tleaf} leaf temperature of each layer (deg C)

\code{ea} vapour pressure (kPa)

\code{gtt} heat conductance to reference height two m above canopy

\code{Rem} emitted radiation (W / m^2)

\code{H} Sensible heat flux from leaf to air (W / m^2)

\code{L} Latent heat of vapourisation from leaf to air (W / m^2)
}
\description{
Calculates leaf temperature from radiation fluxes and reference air
and ground tempertaure
}
\details{
\code{leaftemp} computes the average leaf and air temperature of each canopy layer based on
radiation and evaorative fluxes and reference air and ground temperature. The function
automatically determines whether heat storage should be considered, based the specific heat
capacity of the vegetation layer and the time step of the model.
}
\examples{
# Generate paramaters for function:
tme <- as.POSIXlt(0, origin = "2020-05-04 12:00", tz = "GMT")
previn <- paraminit(20, 10, 10, 15, 2, 80, 11, 500)
vegp <- microctools::habitatvars(3, 50, -5, tme, m = 20)
soilp <- soilinit("Loam")
z<-c((1:20) - 0.5) / 20 * vegp$hgt
# run function (setting conductances in current time step to same as in previous):
ltemp <- leaftemp(11, 80, 101.3, 60, previn$gt, previn$gha, previn$gv, previn$Rabs, previn,
                  vegp, soilp, 0.3)
plot(z ~ ltemp$tleaf, type = "l", xlab = "Leaf temperature", ylab = "Height")
}
