% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{runmodelS}
\alias{runmodelS}
\title{Run model under steady state conditions}
\usage{
runmodelS(climdata, vegp, nmrout, reqhgt, lat, long, metopen = TRUE,
  windhgt = 2, surfwet = 1, groundem = 0.95)
}
\arguments{
\item{climdata}{data.frame of climate variables needed to run the run the model (dataset should follow format of \code{\link[=weather]{weather()}})}

\item{vegp}{a list of vegetation parameters as returned by \code{\link[microctools:habitatvars]{microctools::habitatvars()}}.}

\item{nmrout}{a list of putputs from NicheMapR as returned by \code{\link[=runNMR]{runNMR()}}.}

\item{reqhgt}{height (m) for which microclimate is needed.}

\item{lat}{latitude of location (decimal degrees).}

\item{long}{longitude of location (decimal degrees).}

\item{metopen}{optional logical indicating whether the wind measurement used as an input to
the model is from a nearby weather station located in open ground (TRUE) or above the canopy
for which temperatures are modelled (FALSE - see details)}

\item{windhgt}{height above ground of wind measurement. If \code{metopen} is FALSE, must be above
canopy.}

\item{surfwet}{proportion of leaf surface acting as free water surface}

\item{groundem}{thermal emissivity of ground layer}
}
\value{
a data.frame with the following columns:

\code{obs_time} time of observation. Same as in climdata.

\code{Tref} Temperature (deg C) at reference height as in climdata

\code{Tloc} Temperature (deg C) at height \code{reqhgt}

\code{tleaf} Leaf temperature (deg C) at height \code{reqhgt}. -999 if \code{reqhgt} above canopy
or below ground

\code{RHref} Relative humidity (percentage) at reference height as in climdata.

\code{RHloc} Relative humidity (percentage) at height \code{reqhgt}

\code{RSWloc} Total incoming shortwave radiation at height \code{reghgt} (W/m^2)

\code{RLWloc} Total downward longwave radiation at height \code{reqhgt} (W/m^2)
}
\description{
Rapid method for calculating below or above canopy or below ground microclimate
under steady-state at one user specified height.
}
\details{
This is a rapid implementation of model when time increments are hourly such that
transient heat fluxes and heat storage in canopy can be ignored. Computations are performed
simultaniously on all data, negating need to run in timesteps. Also includes implementation of
model with snow present. Should generally be run using wrapper function \code{\link[=runwithNMR]{runwithNMR()}} but
provided as a standalone function in case data for multiple heights are needed, in whihc case
it can be run multiple times without also running NicheMapR.
}
