% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{runNMR}
\alias{runNMR}
\title{Runs NicheMapR model}
\usage{
runNMR(
  climdata,
  prec,
  lat,
  long,
  Usrhyt,
  Veghyt,
  Refhyt = 2,
  PAI = 3,
  LOR = 1,
  pLAI = 0.8,
  clump = 0,
  REFL = 0.15,
  LREFL = 0.4,
  SLE = 0.95,
  DEP = c(0, 2.5, 5, 10, 15, 20, 30, 50, 100, 200),
  ALTT = 0,
  SLOPE = 0,
  ASPECT = 0,
  ERR = 1.5,
  soiltype = "Loam",
  PE = rep(1.1, 19),
  KS = rep(0.0037, 19),
  BB = rep(4.5, 19),
  BD = rep(1.3, 19),
  DD = rep(2.65, 19),
  cap = 1,
  hori = rep(0, 36),
  maxpool = 1000,
  rainmult = 1,
  SoilMoist_Init = c(0.1, 0.12, 0.15, 0.2, 0.25, 0.3, 0.3, 0.3, 0.3, 0.3)
)
}
\arguments{
\item{climdata}{a data.frame of hourly weather variables in same format as \code{\link[=weather]{weather()}}.
Column obs_time must give times in GMT/UTC.}

\item{prec}{a vector of daily or hourly recipitation values (mm).}

\item{lat}{latitude (decimal degrees, positive in northern hemisphere).}

\item{long}{longitude (decimal degrees, negative west of Greenwich meridian).}

\item{Usrhyt}{Local height (m) at which air temperature, wind speed and humidity are to be computed (see details).}

\item{Veghyt}{At of vegetation canopy (see details).}

\item{Refhyt}{Reference height (m) at which air input climate variables are measured.}

\item{PAI}{single value or vector of hourly or daily values of total plant area per unit ground area of canopy.}

\item{LOR}{Campbell leaf angle distrubution coefficient}

\item{pLAI}{fraction of PAI that is green vegetation.}

\item{clump}{clumpiness factor for canopy (0 to 1, 0 = even)}

\item{REFL}{A single numeric value of ground reflectivity of shortwave radiation.}

\item{LREFL}{A single numeric value of average canopy reflectivity of shortwave radiation.}

\item{SLE}{Thermal emissivity of ground}

\item{DEP}{Soil depths at which calculations are to be made (cm), must be 10 values
starting from 0, and more closely spaced near the surface.}

\item{ALTT}{elevation (m) of location.}

\item{SLOPE}{slope of location (decimal degrees).}

\item{ASPECT}{aspect of location (decimal degrees, 0 = north).}

\item{ERR}{Integrator error tolerance for soil temperature calculations.}

\item{soiltype}{soil type as for \code{\link[=soilparams]{soilparams()}}. Set to NA to use soil parameters
specified below.}

\item{PE}{Air entry potentials (J/kg) (19 values descending through soil for specified soil
nodes in parameter DEP and points half way between). Ignored if soil type is given.}

\item{KS}{Saturated conductivities (kg s/m^3) (19 values descending through soil for
specified soil nodes in parameter DEP and points half way between). Ignored if soil type
is given.}

\item{BB}{Campbell's soil 'b' parameter (19 values descending through soil for specified
soil nodes in parameter DEP and points half way between). Ignored if soil type is given.}

\item{BD}{Soil bulk density (Mg/m^3) (19 values descending through soil for specified
soil nodes in parameter DEP and points half way between). Ignored if soil type is given.}

\item{DD}{Soil density (Mg/m^3) (19 values descending through soil for specified
soil nodes in parameter DEP and points half way between). Ignored if soil type is given.}

\item{cap}{Is organic cap present on soil surface? (1 = Yes, 0 = No).}

\item{hori}{Horizon angles (degrees), from 0 degrees azimuth (north) clockwise in 10 degree intervals.}

\item{maxpool}{Max depth for water pooling on the surface (mm), to account for runoff.}

\item{rainmult}{Rain multiplier for surface soil moisture (used to induce runon).}

\item{SoilMoist_Init}{Initial volumetric soil water content at each soil node (m^3/m^3)}
}
\value{
a list with the following components:

\code{metout} a data.frame of microclimate variables for each hour at height \code{Usrhyt}.

\code{soiltemps} a data.frame of hourly soil temperatures for each depth node.

\code{soilmoist} a data.frame of hourly soil volumetric water content for each depth node.

\code{snowtemp} if snow present, a data,frame of snow temperature (°C), at each of the potential 8 snow layers (see details).
0 if snow not present.

\code{plant} a data.frame of plant transpiration rates (g/m^2/hr), leaf water potentials
(J/kg) and root water potential (J/kg) at each of the 10 specified depths.

\code{nmrout} full NicheMapR output
}
\description{
Wrapper function for running NicheMapR to derive soil moistures
}
\details{
Requires NicheMapR: devtools::install_github('mrke/NicheMapR'). NicheMapR is an integrated
soil moisture and temperature model that treats the vegetation as a single layer (https://mrke.github.io/).
This wrapper function, runs the NicheMapR::microclimate function with reduced parameter
inputs, by specifying sensible default values for other parameters where it is unlikely that these
would be known or explicitely measured at the site. The degree of canopy shading is worked out
explicitely from \code{PAI} at each hourly time interval by estimating canopy tranmission of direct
and diffuse radiation, and by adjusting the amount of radiation that would be absorbed
at the ground surface for a given slope and aspect. Roughness lengths and zero plane displacement
heights are limited to <2m in NicheMapR, so if \code{Veghyt} > 2, it is set to 2m.
This is adequate for computing soil moisture and total evapotranspiration, but will give false
air temperatures for heights below canopy (see \code{\link[=runwithNMR]{runwithNMR()}}. If \code{soiltype} is given, the subsequent
soil parameters are computed from soil type.  If \code{soiltype} is NA, soil paramaters must be specified.
}
\examples{
# Run NicheMapR with default parameters and inbuilt weather datasets
library(NicheMapR)
microout<-runNMR(weather, dailyprecip, 50.2178, -5.32656, 0.05, 0.02, PAI = 1)
# Plot air temperatures
metout <- microout$metout
tmn <- min(metout$TALOC,metout$TAREF)
tmx <- max(metout$TALOC,metout$TAREF)
dday <- metout$DOY+metout$TIME/1440 # Decimal day
plot(metout$TALOC~dday, type="l", col = "red", ylim=c(tmn,tmx), ylab = "Temperature")
par(new = T)
plot(metout$TAREF~dday, type="l", col = "blue", ylim=c(tmn,tmx), ylab = "", xlab = "")
# Plot soil temperatures
soiltemp<-microout$soiltemps
tmn <- min(soiltemp$D0cm,soiltemp$D30cm)
tmx <- max(soiltemp$D0cm,soiltemp$D30cm)
plot(soiltemp$D0cm~dday, type="l", col = "red", ylim=c(tmn,tmx), ylab = "Temperature")
par(new = T)
plot(soiltemp$D30cm~dday, type="l", col = "blue", ylim=c(tmn,tmx), ylab = "", xlab = "")
# Plot soil moistures
soilm <- microout$soilmoist
mmn <- min(soilm$WC2.5cm,soilm$WC30cm)
mmx <- max(soilm$WC2.5cm,soilm$WC30cm)
plot(soilm$WC2.5cm~dday, type="l", col = "red", ylim=c(mmn,mmx), ylab = "Soil moisture")
par(new = T)
plot(soilm$WC30cm~dday, type="l", col = "blue", ylim=c(mmn,mmx), ylab = "", xlab = "")
}
