% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{spinup}
\alias{spinup}
\title{Model spin-up for first time-step}
\usage{
spinup(climdata, vegp, soilp, lat, long, edgedist = 100, reqhgt = NA,
  sdepth = 2, zu = 2, theta = 0.3, thetap = 0.3, merid = 0,
  dst = 0, n = 0.6, plotout = TRUE, steps = 200, metopen = TRUE,
  windhgt = 2, zlafact = 1, surfwet = 1)
}
\arguments{
\item{climdata}{a data.frame of climate variables needed to run the run the model (dataset should follow format of \code{\link[=weather]{weather()}})}

\item{vegp}{a list of vegetation parameters as returned by \code{\link[microctools:habitatvars]{microctools::habitatvars()}}}

\item{soilp}{a list of soil parameters as returned by \code{\link[=soilinit]{soilinit()}}}

\item{lat}{Latitude (decimal degrees)}

\item{long}{Longitude (decimal degrees, negative west of Greenwich meridion)}

\item{edgedist}{distance to open ground (m)}

\item{reqhgt}{optional height for which temperature is required (see details)}

\item{sdepth}{depth of deepest soil node (m)}

\item{zu}{height above ground of reference climate measurements (m)}

\item{theta}{volumetric water content of upper most soil layer in current time step (m^3 / m^3)}

\item{thetap}{volumetric water content of upper most soil layer in previous time step (m^3 / m^3)}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{n}{forward / backward weighting for Thomas algorithm (see \code{\link[=Thomas]{Thomas()}})}

\item{plotout}{optional logical indicating whether to a plot a profile of temperatures
upon completion.}

\item{steps}{number of iterations for which which to run spin-up}

\item{metopen}{optional logical indicating whether the wind measurement used as an input to
the model is from a nearby weather station located in open ground (TRUE) or above the canopy
for which temperatures are modelled (FALSE - see details)}

\item{windhgt}{height above ground of wind measurement. If \code{metopen} is FALSE, must be above
canopy.}

\item{zlafact}{numeric value indicating how close to leaves air temperatures are
needed for (1 - average leaf-air distance, 0.5 = half average leaf-air distance etc.).
Must be greater than 1.}

\item{surfwet}{fraction of vegetation surface acting like a free-water surface}
}
\value{
a list of model outputs as for \code{\link[=paraminit]{paraminit()}} or \code{\link[=runonestep]{runonestep()}}
}
\description{
\code{spinup} runs the model repeatedly using data form the first time-step
for a set number of steps to ensure initial conditions are stable and
appropriate soil temperatures are set.
}
\details{
If \code{reqhgt} is set, and below the height of the canopy, the canopy node nearest
to that height is set at the value specified. The returned value \code{tabove} is then the
temperature at the top of the canopy. If \code{reqhgt} is above canopy, nodes are calculated
automatically, but \code{tabove} is the temperature at height \code{reqhgt}. If \code{reqhgt} is
negative, the soil node nearest to that height is set at the value specified.
}
\examples{
tme<-as.POSIXlt(weather$obs_time, format = "\%Y-\%m-\%d \%H:\%M", tz = "UTC")
vegp <- habitatvars(4, 50, -5, tme, m = 20)
soilp<- soilinit("Loam")
# run spinup and produce profile plot
modelout<-spinup(weather, vegp, soilp, lat = 50, long = -5)
}
