% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ThomasV}
\alias{ThomasV}
\title{Thomas algorithm for solving simultanious vapour fluxes}
\usage{
ThomasV(
  Vo,
  tn,
  pk,
  theta,
  thetap,
  relhum,
  tair,
  tsoil,
  zth,
  gt,
  Vflux,
  f = 0.6,
  previn,
  soilp
)
}
\arguments{
\item{Vo}{a vector of air vapour concentrations for each canopy node in the previos timestep (mol fraction)}

\item{tn}{vector of air temperatures (deg C) for each canopy node in the current timestep (deg C)}

\item{pk}{atmospheric pressure (kPa)}

\item{theta}{Volumetric water content of the upper most soil layer in the current time step (m3 / m3)}

\item{thetap}{Volumetric water content of the upper most soil layer in the previous time step (m3 / m3)}

\item{relhum}{relative humidity (percentage) at reference height 2 m above canopy in current time step (percentage)}

\item{tair}{air temperature at reference height 2 m above canopy in current time step (deg C)}

\item{tsoil}{temperature of upper soil layer in current time step (deg C)}

\item{zth}{heightdifference between each canopy node and that directly below it. the first value is
the height difference between the lowest canopy node and the ground}

\item{gt}{vector of molar conductances between each canopy node at that directly below it (mol / m^2 / sec).
The first value is the conductivity between the ground and the lowest node, and the last value the
conductivity between the highest node and reference height.}

\item{Vflux}{Total vapour flux from leaves to air (mol /m^3)}

\item{f}{forward / backward weighting of algorithm (as for \code{\link[=Thomas]{Thomas()}})}

\item{previn}{a list of model outputs form the previous timestep}

\item{soilp}{a list of soil parameters as returned by \code{\link[=soilinit]{soilinit()}}}
}
\value{
a vector of vapour concentrations expressed as mole fractions for each canopy node in the
current time step. The first value is that for the ground and the last value that at reference height
}
\description{
\code{ThomasV} implements the Thomas algorithm for solving simultanious vapour
fluxes between air layers.
}
\seealso{
\code{\link[=Thomas]{Thomas()}}
}
