% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheMapRcode.R
\name{tleafS}
\alias{tleafS}
\title{Steady-state leaf and air temperature}
\usage{
tleafS(
  tair,
  tground,
  relhum,
  pk,
  theta,
  gtt,
  gt0,
  gha,
  gv,
  gL,
  Rabs,
  vegem,
  soilb,
  Psie,
  Smax,
  surfwet,
  leafdens
)
}
\arguments{
\item{tair}{air temperature at reference height (deg C)}

\item{tground}{ground surface temperature (as returned by e.g. \code{\link[=runNMR]{runNMR()}})}

\item{relhum}{relative humidity at reference height (percentage)}

\item{pk}{atmospheric pressure (kPa)}

\item{theta}{volumetric water content of upper most soil layer in current time step (m^3 / m^3)
as returned by e.g. \code{\link[=runNMR]{runNMR()}}}

\item{gtt}{molar conductivity from leaf height to reference height as returned by \code{\link[=gturb]{gturb()}}
and \code{\link[=gcanopy]{gcanopy()}}}

\item{gt0}{molar conductivity from leaf height to ground as returned by \code{\link[=gcanopy]{gcanopy()}}}

\item{gha}{boundary layer conductance of leaf as returned by \code{\link[=gforcedfree]{gforcedfree()}}}

\item{gv}{combined boundary layer and stomatal conductance of leaf}

\item{gL}{combined boundary layer and leaf-air conductance given by 1/(1/gha+1/(uz*ph))}

\item{Rabs}{radiation absorbed by leaf as returned by e.g. \code{\link[=leafabs]{leafabs()}}}

\item{soilb}{Shape parameter for Campbell soil model (dimensionless, > 1) as returned by
\code{\link[=soilinit]{soilinit()}}}

\item{Psie}{Soil matric potential (J / m^3) as returned by \code{\link[=soilinit]{soilinit()}}}

\item{Smax}{Volumetric water content at saturation (m3 / m3)}

\item{surfwet}{proportion of leaf surface acting as free water surface}

\item{leafdens}{Total one sided leaf area per m^3 at desired height}
}
\value{
a list of the following:

\code{tleaf} leaf temperature

\code{tn} air temperature

\code{rh} relative humidity
}
\description{
Rapid method for calculating steady-state leaf and air temperature at one height
}
