% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{A data frame of hourly weather}
\format{a data frame with the following elements:
\describe{
\item{obs_time}{POSIXlt object of dates and times}
\item{temp}{temperature (degrees C)}
\item{relhum}{relative humidity (percentage)}
\item{pres}{atmospheric press (kPa)}
\item{swrad}{Total incoming shortwave radiation (W / m^2)}
\item{difrad}{Diffuse radiation (W / m^2)}
\item{skyem}{Sky emissivity (0-1)}
\item{windspeed}{Wind speed (m/s)}
\item{winddir}{Wind direction (decimal degrees)}
}}
\usage{
weather
}
\description{
A data frame of hourly weather in 1995 at Camborne weather station, Cornwall (50.2178N, 5.32656W)
}
\keyword{datasets}
