% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{windprofile}
\alias{windprofile}
\title{Calculates wind profile for entire canopy}
\usage{
windprofile(ui, zi, zo, a, PAI, hgt, psi_m = 0, hgtg = 0.05 * hgt,
  zm0 = 0.004)
}
\arguments{
\item{ui}{wind at reference height (m / s)}

\item{zi}{height of wind measurement (m)}

\item{zo}{height for which wind is required (m)}

\item{a}{attenuation coefficient as returned by \code{\link[microctools:attencoef]{microctools::attencoef()}}}

\item{PAI}{plant area index (m / m)}

\item{hgt}{height of canopy (m)}

\item{psi_m}{diabatic correction factor as return by \code{\link[microctools:diabatic_cor]{microctools::diabatic_cor()}}}

\item{hgtg}{height of ground vegetation layer below canopy (m)}

\item{zm0}{roughness length (m) of vegetation layer below canopy as returned by \code{\link[microctools:roughlength]{microctools::roughlength()}}}
}
\value{
wind speed (m /s) at height \code{zo}
}
\description{
calculates wind speed at any given point above or below canopy
from wind speed at reference height
}
\examples{
zo<- c(0:200) / 100
uz <- windprofile(2, 2, zo, 2.3, 3, 2)
plot(zo ~ uz, type = "l", xlab = "wind speed", ylab = "height")
}
